+++ cal-heatmap.js
@@ -1,4 +1,4 @@
-/*! cal-heatmap v2.0.3 (Thu Mar 21 2013 13:19:49)
+/*! cal-heatmap v2.1.0 (Thu Mar 28 2013 20:01:04)
  *  ---------------------------------------------
  *  A module to create calendar heat map to visualise time data series a la github contribution graph
  *  https://github.com/kamisama/cal-heatmap
@@ -17,8 +17,9 @@ var CalHeatMap = function() {
 		// DOM ID of the container to append the graph to
 		id : "cal-heatmap",
 
-		// Threshold for each scale
-		scale : [10,20,30,40],
+		// ================================================
+		// DOMAIN
+		// ================================================
 
 		// Number of domain to display on the graph
 		range : 12,
@@ -31,24 +32,9 @@ var CalHeatMap = function() {
 
 		domainGutter : 2,
 
-		format : {
-			// Formatting of the date when hovering an subdomain block
-			// @default : null, will use the formatting according to domain type
-			date : null,
+		domain : "hour",
 
-			// Formatting of domain label
-			// @default : null, will use the formatting according to domain type
-			legend : null
-		},
-
-		// Callback when clicking on a time block
-		onClick : null,
-
-		// Whether to display the scale
-		displayScale : true,
-
-		// Name of the items to represent in the calendar
-		itemName : ["item", "items"],
+		subDomain : "min",
 
 		// Start date of the graph
 		// @default now
@@ -61,9 +47,47 @@ var CalHeatMap = function() {
 		// When false, the calendar will be left empty
 		loadOnInit : true,
 
-		domain : "hour",
+		// ================================================
+		// SCALE
+		// ================================================
 
-		subDomain : "min",
+		// Threshold for the scale
+		scale : [10,20,30,40],
+
+		// Whether to display the scale
+		displayScale : true,
+
+		// ================================================
+		// TEXT FORMATTING
+		// ================================================
+
+		format : {
+			// Formatting of the date when hovering an subdomain block
+			// @default : null, will use the formatting according to domain type
+			date : null,
+
+			// Formatting of domain label
+			// @default : null, will use the formatting according to domain type
+			legend : null
+		},
+
+		// Name of the items to represent in the calendar
+		itemName : ["item", "items"],
+
+		cellLabel : {
+			empty: "{date}",
+			filled: "{count} {name} {connector} {date}"
+		},
+
+		scaleLabel : {
+			lower: "less than {min} {name}",
+			inner: "between {down} and {up} {name}",
+			upper: "more than {max} {name}"
+		},
+
+		// ================================================
+		// BROWSING
+		// ================================================
 
 		// Animation duration
 		duration : 500,
@@ -77,6 +101,16 @@ var CalHeatMap = function() {
 			nextLabel : "Next",
 			previousLabel : "Previous"
 		},
+
+		// ================================================
+		// CALLBACK
+		// ================================================
+
+		// Callback when clicking on a time block
+		onClick : null,
+
+		// Callback when clicking on a time block
+		afterLoad : null,
 
 		// Callback after loading the next domain in the calendar
 		afterLoadNextDomain : function(start) {},
@@ -268,7 +302,9 @@ var CalHeatMap = function() {
 		self.paint();
 
 
-
+		if (self.options.afterLoad !== null) {
+			self.afterLoad();
+		}
 
 		// Display scale if needed
 		if (self.options.displayScale) {
@@ -549,7 +585,24 @@ CalHeatMap.prototype = {
 	 * @param  int		itemNb	Number of items in that date
 	 */
 	onClick : function(d, itemNb) {
-		return this.options.onClick(d, itemNb);
+		if (typeof (this.options.onClick) === "function") {
+			return this.options.onClick(d, itemNb);
+		} else {
+			console.log("Provided callback for onClick is not a function.");
+			return false;
+		}
+	},
+
+	/**
+	 * Callback to fire after drawing the calendar, but before filling it
+	 */
+	afterLoad : function() {
+		if (typeof (this.options.afterLoad) === "function") {
+			return this.options.afterLoad();
+		} else {
+			console.log("Provided callback for afterLoad is not a function.");
+			return false;
+		}
 	},
 
 	/**
@@ -558,8 +611,13 @@ CalHeatMap.prototype = {
 	 * @param  Date		end		Domain end date
 	 */
 	afterLoadPreviousDomain: function(start) {
-		var subDomain = this.getSubDomain(start);
-		return this.options.afterLoadPreviousDomain(subDomain.shift(), subDomain.pop());
+		if (typeof (this.options.afterLoadPreviousDomain) === "function") { console.log(typeof (this.options.afterLoadPreviousDomain));
+			var subDomain = this.getSubDomain(start);
+			return this.options.afterLoadPreviousDomain(subDomain.shift(), subDomain.pop());
+		} else {
+			console.log("Provided callback for afterLoadPreviousDomain is not a function.");
+			return false;
+		}
 	},
 
 	/**
@@ -568,8 +626,13 @@ CalHeatMap.prototype = {
 	 * @param  Date		end		Domain end date
 	 */
 	afterLoadNextDomain: function(start) {
-		var subDomain = this.getSubDomain(start);
-		return this.options.afterLoadNextDomain(subDomain.shift(), subDomain.pop());
+		if (typeof (this.options.afterLoadNextDomain) === "function") {
+			var subDomain = this.getSubDomain(start);
+			return this.options.afterLoadNextDomain(subDomain.shift(), subDomain.pop());
+		} else {
+			console.log("Provided callback for afterLoadNextDomain is not a function.");
+			return false;
+		}
 	},
 
 	formatNumber: d3.format(",g"),
@@ -605,11 +668,18 @@ CalHeatMap.prototype = {
 			.text(function(d) {
 				var nextThreshold = parent.options.scale[d+1];
 				if (d === 0) {
-					return "less than " + parent.options.scale[d] + " " + parent.options.itemName[1];
+					return (parent.options.scaleLabel.lower).format({
+						min: parent.options.scale[d],
+						name: parent.options.itemName[1]});
 				} else if (d === parent.options.scale.length) {
-					return "more than " + parent.options.scale[d-1] + " " + parent.options.itemName[1];
+					return (parent.options.scaleLabel.upper).format({
+						max: parent.options.scale[d-1],
+						name: parent.options.itemName[1]});
 				} else {
-					return "between " + parent.options.scale[d-1] + " and " + parent.options.scale[d] + " " + parent.options.itemName[1];
+					return (parent.options.scaleLabel.inner).format({
+						down: parent.options.scale[d-1],
+						up: parent.options.scale[d],
+						name: parent.options.itemName[1]});
 				}
 			})
 		;
@@ -658,10 +728,17 @@ CalHeatMap.prototype = {
 						var subDomainUnit = parent._domainType[parent.options.subDomain].extractUnit(d);
 
 						return (
-						(data[domainUnit].hasOwnProperty(subDomainUnit) ?
-							(parent.formatNumber(data[domainUnit][subDomainUnit]) + " " + parent.options.itemName[(data[domainUnit][subDomainUnit] > 1 ? 1 : 0)] + " " + parent._domainType[parent.options.subDomain].format.connector + " ") :
-							""
-							) + parent.formatDate(d));
+						data[domainUnit].hasOwnProperty(subDomainUnit) ?
+							(parent.options.cellLabel.filled).format({
+								count: parent.formatNumber(data[domainUnit][subDomainUnit]),
+								name: parent.options.itemName[(data[domainUnit][subDomainUnit] > 1 ? 1 : 0)],
+								connector: parent._domainType[parent.options.subDomain].format.connector,
+								date: parent.formatDate(d)
+							}) :
+							(parent.options.cellLabel.empty).format({
+								date: parent.formatDate(d)
+							})
+						);
 					});
 				}
 			}
@@ -1012,6 +1089,19 @@ CalHeatMap.prototype = {
 	}
 };
 
+/**
+ * Sprintf like function
+ * @source http://stackoverflow.com/a/4795914/805649
+ * @return String
+ */
+String.prototype.format = function () {
+	var formatted = this;
+	for (var prop in arguments[0]) {
+		var regexp = new RegExp("\\{" + prop + "\\}", "gi");
+		formatted = formatted.replace(regexp, arguments[0][prop]);
+	}
+	return formatted;
+};
 
 /**
  * AMD Loader